+++ theme/template/functions/consts.php
@@ -3,18 +3,22 @@
  * Category names.
  */
 define('NEWS_CATEGORY', 'news');
-define('SPONSOR_PLATINUM', 'platinum-sponsors');
-define('SPONSOR_GOLD', 'gold-sponsors');
-define('SPONSOR_SILVER', 'silver-sponsors');
-define('SPONSOR_OPERATION', 'operation-sponsors');
+define('SUPPORTER_PLATINUM', 'platinum-supporters');
+define('SUPPORTER_GOLD', 'gold-supporters');
+define('SUPPORTER_SILVER', 'silver-supporters');
+define('SUPPORTER_OPERATION', 'operation-supporters');
 
 /**
  * Page names.
  */
-define('SPONSORS_PAGE', 'sponsors');
+define('SUPPORTERS_PAGE', 'supporters');
 define('TIMETABLE_PAGE', 'timetable');
 
 /**
  * Custom post types.
  */
 define('SESSION_POST_TYPE', 'session');
+define('SUPPORTER_PLATINUM_POST_TYPE', 'platinum-supporter');
+define('SUPPORTER_GOLD_POST_TYPE', 'gold-supporter');
+define('SUPPORTER_SILVER_POST_TYPE', 'silver-supporter');
+define('SUPPORTER_OPERATION_POST_TYPE', 'operation-supporter');