+++ theme/template/single.php
@@ -4,13 +4,11 @@ switch (get_post_type($post->ID)) {
     include_once get_template_directory() . '/posts/session.php';
     break;
 
-  // TODO: might cause high load when post grows.
-  // Consider to refactor to get the post once, then match with the categories.
-  case in_category(SPONSOR_PLATINUM, $post->ID):
-  case in_category(SPONSOR_GOLD, $post->ID):
-  case in_category(SPONSOR_SILVER, $post->ID):
-  case in_category(SPONSOR_OPERATION, $post->ID):
-    include_once get_template_directory() . '/posts/sponsor.php';
+  case SUPPORTER_PLATINUM_POST_TYPE:
+  case SUPPORTER_GOLD_POST_TYPE:
+  case SUPPORTER_SILVER_POST_TYPE:
+  case SUPPORTER_OPERATION_POST_TYPE:
+    include_once get_template_directory() . '/posts/supporter.php';
     break;
 
   case in_category(NEWS_CATEGORY, $post->ID):